Implementation ioc registrations

+++ src/Fiap.SoftwareEngineering.Netflix.Domain/Services/DomainWriterService.cs
@@ -1,25 +1,40 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-using Fiap.SoftwareEngineering.Netflix.Domain.Abstractions;
-using Fiap.SoftwareEngineering.Netflix.Domain.Abstractions.Services;
+﻿using Fiap.SoftwareEngineering.Netflix.Domain.Abstractions.Services;
 using Fiap.SoftwareEngineering.Netflix.Domain.Abstractions.Validations;
+using Fiap.SoftwareEngineering.Netflix.Domain.Validations;
 using Fiap.SoftwareEngineering.Netflix.Repository.Abstractions;
 using Fiap.SoftwareEngineering.Netflix.Validation;
+using System;
+using System.Collections.Generic;
+using System.Runtime.CompilerServices;
+using System.Threading;
+using System.Threading.Tasks;
+using Fiap.SoftwareEngineering.Netflix.Domain.Abstractions.Entities;
+using Fiap.SoftwareEngineering.Netflix.Repository.Abstractions.Entities;
 
 namespace Fiap.SoftwareEngineering.Netflix.Domain.Services
 {
-    public class DomainWriterService<TEntity> : IDomainWriterService<TEntity>
-        where TEntity : class
+    public class DomainWriterService<TEntity> : DomainWriterService<TEntity, int>, IDomainWriterService<TEntity>
+        where TEntity : class, IEntity<int>
     {
-        protected IRepositoryWriter<TEntity> Repository;
-        protected IDomainValidator<TEntity> Validator;
+        public DomainWriterService(IRepositoryWriter<TEntity> repository, IDomainValidator<TEntity> validator, IDomainReaderService<TEntity> domainReaderService)
+            : base(repository, validator, domainReaderService)
+        {
+        }
+    }
 
-        public DomainWriterService(IRepositoryWriter<TEntity> repository, IDomainValidator<TEntity> validator)
+    public class DomainWriterService<TEntity, TKey> : IDomainWriterService<TEntity, TKey>
+        where TEntity : class, IEntity<TKey>
+        where TKey : IEquatable<TKey>
+    {
+        protected readonly IRepositoryWriter<TEntity> Repository;
+        protected readonly IDomainValidator<TEntity> Validator;
+        protected readonly IDomainReaderService<TEntity, TKey> DomainReaderService;
+
+        public DomainWriterService(IRepositoryWriter<TEntity> repository, IDomainValidator<TEntity> validator, IDomainReaderService<TEntity, TKey> domainReaderService)
         {
             Repository = repository ?? throw new ArgumentNullException(nameof(repository));
             Validator = validator ?? throw new ArgumentNullException(nameof(validator));
+            DomainReaderService = domainReaderService ?? throw new ArgumentNullException(nameof(domainReaderService));
         }
 
         public async Task<TEntity> AddAsync(TEntity entity, CancellationToken cancellationToken = default)
@@ -35,8 +50,8 @@ namespace Fiap.SoftwareEngineering.Netflix.Domain.Services
             var operationResult = await Repository.UnitOfWork.SaveChangesAsync(cancellationToken);
             if (operationResult > 0) return entityInserted;
             
-            var notification = new Notification("InvalidDatabaseOperation", "Invalid Database Operation.");
-            Validator.NotificationContext.AddNotification(notification);
+            AddNotiticationMessage(nameof(StaticMessages.InvalidDatabaseOperation),
+                StaticMessages.InvalidDatabaseOperation);
             return entityInserted;
         }
 
@@ -61,12 +76,15 @@ namespace Fiap.SoftwareEngineering.Netflix.Domain.Services
             if (!isValid)
                 return;
 
+            if (!await Exists(entity.Key, cancellationToken))
+                return;
+
             await Repository.UpdateAsync(entity, cancellationToken);
             var operationResult = await Repository.UnitOfWork.SaveChangesAsync(cancellationToken);
             if (operationResult > 0) return;
             
-            var notification = new Notification("InvalidDatabaseOperation", "Invalid Database Operation.");
-            Validator.NotificationContext.AddNotification(notification);
+            AddNotiticationMessage(nameof(StaticMessages.InvalidDatabaseOperation),
+                StaticMessages.InvalidDatabaseOperation);
         }
 
         public async Task UpdateRangeAsync(IEnumerable<TEntity> entities, CancellationToken cancellationToken = default)
@@ -74,11 +92,15 @@ namespace Fiap.SoftwareEngineering.Netflix.Domain.Services
             if (entities == null)
                 throw new ArgumentNullException(nameof(entities));
 
-            await Repository.UnitOfWork.ExecuteInTransactionAsync(async transaction =>
-            {
-                foreach (var entity in entities)
-                    await UpdateAsync(entity, cancellationToken);
-            }, cancellationToken);
+            if (!await IsValid(entities, cancellationToken))
+                return;
+
+            await Repository.UpdateRangeAsync(entities, cancellationToken);
+            var operationResult = await Repository.UnitOfWork.SaveChangesAsync(cancellationToken);
+            if (operationResult > 0) return;
+            
+            AddNotiticationMessage(nameof(StaticMessages.InvalidDatabaseOperation),
+                StaticMessages.InvalidDatabaseOperation);
         }
 
         public async Task RemoveAsync(TEntity entity, CancellationToken cancellationToken = default)
@@ -86,12 +108,15 @@ namespace Fiap.SoftwareEngineering.Netflix.Domain.Services
             if (entity == null)
                 throw new ArgumentNullException(nameof(entity));
 
+            if (!await Exists(entity.Key, cancellationToken))
+                return;
+
             await Repository.RemoveAsync(entity, cancellationToken);
             var operationResult = await Repository.UnitOfWork.SaveChangesAsync(cancellationToken);
             if (operationResult > 0) return;
             
-            var notification = new Notification("InvalidDatabaseOperation", "Invalid Database Operation.");
-            Validator.NotificationContext.AddNotification(notification);
+            AddNotiticationMessage(nameof(StaticMessages.InvalidDatabaseOperation),
+                StaticMessages.InvalidDatabaseOperation);
         }
 
         public async Task RemoveRangeAsync(IEnumerable<TEntity> entities, CancellationToken cancellationToken = default)
@@ -99,11 +124,40 @@ namespace Fiap.SoftwareEngineering.Netflix.Domain.Services
             if (entities == null)
                 throw new ArgumentNullException(nameof(entities));
 
-            await Repository.UnitOfWork.ExecuteInTransactionAsync(async transaction =>
-            {
-                foreach (var entity in entities)
-                    await RemoveAsync(entity, cancellationToken);
-            }, cancellationToken);
+            if (!await IsValid(entities, cancellationToken))
+                return;
+
+            await Repository.RemoveRangeAsync(entities, cancellationToken);
+            var operationResult = await Repository.UnitOfWork.SaveChangesAsync(cancellationToken);
+            if (operationResult > 0) return;
+            
+            AddNotiticationMessage(nameof(StaticMessages.InvalidDatabaseOperation),
+                StaticMessages.InvalidDatabaseOperation);
+        }
+
+        private async Task<bool> IsValid(IEnumerable<TEntity> entities, CancellationToken cancellationToken = default)
+        {
+            var isValid = false;
+
+            foreach (var entity in entities)
+                isValid = await Validator.ValidateDomainAsync(entity, cancellationToken);
+
+            return isValid;
+        }
+
+        private void AddNotiticationMessage(string key, string message)
+        {
+            var notification = new Notification(key, message);
+            Validator.NotificationContext.AddNotification(notification);
+        }
+
+        private async Task<bool> Exists(TKey key, CancellationToken cancellationToken = default)
+        {
+            var exists = await DomainReaderService.Exists(key, cancellationToken);
+            if (!exists)
+                AddNotiticationMessage(nameof(StaticMessages.EntityNotFound), StaticMessages.EntityNotFound);
+
+            return exists;
         }
     }
 }
